Python/Integration/requirements.txt
requests>=2.31.0
packaging>=23.1
watchdog>=3.0.0
PyYAML>=6.0
pytest>=7.4.0
pdoc3>=0.10.0
